Backend/requirements.txt
# Backend Dependencies
fastapi
uvicorn[standard]
httpx
python-dotenv
ddgs
requests
google-play-scraper
rapidfuzz
playwright
